Updated to use new forward
Fixed a bug that allowed players to teleport to any jump even if they have not been to it by using the parameterized version of the command.

Specifying a 0 as a jump number in the parameterized command (i.e. /goto <jump> 0) now teleports to the end of the jump (progress dependent).